requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
websockets==13.1
